Added a new property for the ADUser Object
I have added the whenCreated property to the ADUser class as this may be
something that I will use later on in the development of the
application. It'll be used primarily to give the users of the
application an idea of when the account was created (i.e. recently, or
far in the past).