I have started to integrate a DB for this application. At the time of this commit I am using the entity framework to enter basic information about users that are now being created thru the application. the thinking of what I've built here is as follows:
The application will be used to keep track of new users that are created
in the domain. We only care about basic information about the user being
created such as the username, first, middle and last name. We are also
recording who is creating the newly created user ... this is an
important part of what I am trying to solve so that we know who created
what account and when.

What you see here is just the very basic starting point of what I
envision this application to do. As I continue working on this, no
doubt the information that I am currently storing will change!