removed the old migrations folder and enabled migrations once more on this project (I really need to sit down and look at how I can prevent doing this on my main machine and laptop as I work on this project). I then re-enabled migrations and pushed out the changes to the db.
The other change you see in this commit has to do with the way that
accounts that don't have an entry in the DomainUsers table are handled.
When a user views an account that doesn't exist in the DomainUsers
table, it would store the username in the DomainUsers table in the
created by field. Instead of storing the username, I am now storing the
first + last name of the user instead. This now makes it very easy to
tell who created the account!