Made a change on how accounts that are created outside the application are recorded in the application. In the past, when a user who is logged into the application browses to a user and views him/her and this user was created outside the application an entry was created in the DomainUsers table with the logged in user as the user who created the entry in the table. This doesn't make any sense because the account was not actually created by the user. I made a change so that instead of logging in the username, I am now storing the string 'System Generated' in the DomainUsers table and also the entry stored in the UserUpdateHistory that is made when this scenario happens.
The other change I made in this commit pertains to how the update
history is shown to the user. This is a bit of a hack at the moment but
I am now showing two columns on the table with the first column showing
information such as when the update occured, what type of update it is
and who made the update. The second column is just the details of what
has changed. This may also change in the future as it's a bit of a hack
as well (nesting tables inside of other tables just leaves a bad taste
in my mouth).