Added a new EF entity to store active directory settings. I didn't like the fact that this was stored in the web.config file and now that these values are stored in the database they will be easier to edit/change as needed. I removed the old values that were stored in the web.config file and tested this out and it's working as expected.
One other change made in this commit was I had to change how the
description of each group is retrieved and shown to the user when they
are viewing the group membership of the user. I was using the 'notes'
property to show this in the past, but now I am using the 'info'
property as this is what's used when creating and editing a group in the
application.